+++ .env.example
@@ -1,1 +1,2 @@
-VITE_API_BASE_URL=https://express-jwt-auth-template.vercel.app/api
+# VITE_API_BASE_URL=https://express-jwt-auth-template.vercel.app/api
+VITE_API_BASE_URL=http://13.209.24.80:8080
